cli-node: refactor git into GitUtils
