+++ packages/cli-node/src/index.ts
@@ -19,3 +19,5 @@
  *
  * @packageDocumentation
  */
+
+export * from './git';
